requirements.txt
# Django core
Django>=6.0

# Database
psycopg2-binary>=2.9.9

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
django-filter>=25.0

# Payment processing
stripe>=7.0.0
paypalrestsdk>=1.13.1

# Production server
gunicorn>=21.0.0
whitenoise>=6.0.0

# Image handling
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0
